Set a maximum length on topic display
This change introduces the gr-limited-text element, which, given a
string of text and a numeric limit, will display the text up to that
limit. If the length of the text exceeds that limit, an ellipsis is
added to indicate the truncation and a tooltip is enabled to display the
full string.

Displaying the topic string is limited to a reasonable number of
characters in this way in both the change metadata and the change list.

Bug: Issue 7197
Change-Id: If826fc2f2657e8ed12275e65f25f2d40c3180838

Former-commit-id: 8553bf460e8e3a234f6399c7252d2c0342a0c526