Give change metadata chips a disabled state
Previously, there was no indication that any progress was being made
when a vote, hashtag, or topic was deleted from the change metadata.

This change gives gr-account-chip and gr-linked-chip a disabled property
that reflects to an attribute, and uses CSS to disable pointer events
and change the opacity of the content.

Bug: Issue 6688
Change-Id: I2a9e61b33ad30af0c5f0b59d01769f3aae8ec9ea

Former-commit-id: f3c0a367f507013666855da8c094e388b106f871