ChangeIT: remove unnecessary reviewer addition in mute test
The point of the test is to confirm that the reviewer (in this case
"user") still receives email notifications when the change is muted.

It is unnecessary to also add the change owner as reviewer. This was
a copy and paste mistake when adding the test based on the test of
the ignore method.

Change-Id: I2992400cb5da1b1aed98486bdd79774fd886457e

Former-commit-id: dda0082f3541aac0b4c28edfb22114b0b7e8f01e