Use injector and factory to instantiate SubmitRuleEvaluator
All callers of SubmitRuleEvaluator get AccountCache, Accounts and
Emails injected just to pass them to SubmitRuleEvaluator manually.

This commit adds a Factory to SubmitRuleEvaluator and migrates callers
to use the factory instead to remove boilerplate.

In addition, this commit moves the instantiation of ChangeControl from
the constructor closer to where it is needed to prevent the constructor
from throwing an OrmException.

Change-Id: I429e7a67e4f1a45be1cf2c2f05692661c086cabe

Former-commit-id: 6942c1ba48d2027625f33cf75f80c2de7f02eb23