Fix _readResponsePayload runtime errors
Before I6398b74779, the REST API interface's `getResponseObject` did not
make use of any other methods, and thus did not need to be invoked with
a `this` context. When this changed, the handful of locations that
invoke `getResponseObject` without binding were broken silently,
resulting in runtime exceptions where the app fails to parse responses
to some network actions. For example, when deleting a topic.

With this change, the invocations use corrected bindings and tests are
added to confirm that runtime exceptions do not occur.

Change-Id: I682d3d5f5fc4aaec03762e17d32afec084bbdebf

Former-commit-id: 0f3874b3cfba0d0b78edc6bdedd52bd873e72cbd