Fix bug in Send button disable state
The object that tracks the existence of drafts was erroneously treated
as an array.

This change calls Object.keys on the drafts object before checking its
length.

In addition, the watcher that computed the Send button disable state did
not check for the state of the _includeComments flag -- now, it does.

Change-Id: I061053715b87de51d291503a3699550c0ed46020

Former-commit-id: 2eed3c31ad5d107593334f767d3e4a909eb1dfa8