Don't load ChangeNotes when lazyload is false
When we serve a change query and format the response in ChangeJson,
lazyload is disabled. This means that ChangeData#notes() throws an
OrmException.

When specifying DETAILED_LABELS we were calling #notes() irrespective of
lazyload resulting in an OrmException for the change and omitting it
from the results. This was only occuring for merged changes.

This commit fixes the bug by constructing ChangeNotes from the index
data instead if lazyload is false and adds a test for all change
options.

Change-Id: Id08ef0e6fa71a875acdfdb7d4a669af2ba38a0cc

Former-commit-id: 86fa9612ec30f79e7616a744c1a2004e60386bf9