Swap 'expand inline' and 'reviewed' controls
In the file list, the 'expand inline' and 'reviewed' controls should be
swapped. Contextually, it makes more sense for the 'expand' control to
appear first, as one needs no information about the file in order to
want to expand it -- in the case of the reviewed checkbox, the user is
expected to have viewed the file before toggling the checkbox state.

In addition, some work was done to migrate the reviewed checkbox to the
new button-like styling designed by Arnab.

Bug: Issue 7275
Change-Id: I5a052da8fdce79523709e9c98b01466cb26eb242

Former-commit-id: 355b6b9914cc54b163e5ea280fac235488d3b57f