DisablePlugin: Explicitly check for capability in apply()
The class is annotated as @RequiresCapability(ADMINISTRATE_SERVER), but
it is only effective when invoked by REST.

If the endpoint is invoked from the internal API, the capability check
is not done.

Replace the annotation with an explicit capability check in the apply()
method.

Also add a test for calling the endpoint from the internal API.

Change-Id: Ie109c6a5d291d4ffcf8563e06a25b8073b937b0c

Former-commit-id: e42f77645aafd3f1663268ad99c6897a03181795