On account creation evict account from account cache
On account creation the account may be indexed several times (e.g. after
addition of external IDs or SSH keys). Reindex is done by evicting the
account from the cache. The first evict call to the cache doesn't find
the account since it's not created yet. This causes the account cache to
cache a missing account instance for the new account ID. When the
account creation is done we must evict this missing account instance
from the cache.

Bug: Issue 7273
Change-Id: Ifa2061d5e83921f45715ba64ab69d2a1e744e5a1
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: c53d29c592534a7728dba881ac2fcacdfe807f49