Require lazyload when calling createCommitMessageOnSubmit
In ChangeJson we previously cheated around lazyload: When obtaining a
ChangeControl or ChangeNotes using the indirection of
ChangeControl#getNotes() we didn't respect lazyload.

With the PermissionBackend refactoring and the deprecation of
ChangeControl, this changed as we always get ChangeNotes using
ChangeData#notes() which fails with an OrmException if lazyload = false.

This commit adds options that require notes. These are not critical for
performance, because in all of the cases we open the repository manually
in ChangeJson to read commit details. Reading notes then is less of an
effort.

Change-Id: Icc2e82cd38cf351f9c6c742c2e51e341a9d44c41

Former-commit-id: 66b8522ed6ff38f9dfe68299e5036c8ced5c9e49