Make bzl files compatible with mutating behavior of += on lists
In the future versions of Bazel (starting with 0.7 or 0.8) += on lists
will behave like in Python, by mutating the left hand side list. If this
list is frozen it will cause a runtime error. To preserve the previous
behavior such lists should be either copied or the `a = a + b` syntax
should be used (which is completely equivalent with the old behavior of
+= on lists).

Change-Id: I3ef62e0c1f3b5ceca705ed4fa331122c3190991e

Former-commit-id: 8f5ca5abf9a70ba10e16b94f9e5cfbbcdaff6c04