Add ProjectPermission.READ_NO_CONFIG
VisibleRefsFilter has a performance short-circuit that returns all refs
minus refs/meta/config. This is needed in larger setups as the other
checks put too much load on the CPU.

As we are migrating to PermissionBackend, this can be a standalone
permission that each permission backend can handle depending on the
performance implications there.

This commit adds ProjectPermission.READ_NO_CONFIG and makes
ProjectControl#allRefsAreVisible() private.

Change-Id: I38520cffefaa2bafa2f7d4ba3bc6e3a641d866e3

Former-commit-id: 5ae5959b6c2e21776ebec0009746adf0037fc802