Remove "My Drafts" menu items from all user preferences
Drafts no longer exist, so these menus are obsolete. The migration only
touches items that exactly match the default query; it's not clear
what we would do in the general case of a query containing the draft at
an arbitrary point in the query tree. Fortunately [is:draft] is now a
predicate that matches no changes, so the menu items that we don't touch
will continue to work and have somewhat reasonable behavior.

Change-Id: Ib3baab6c1da8b0c38449ae95fe6bb092a1ca826c

Former-commit-id: 5124fab8052265a891a7695de47eafe60c49a22a