Add and test failure cases in gr-change-metadata's _computeValueTooltip
When navigating from change to change, gr-change-metadata must re-render
the list of labels and votes in those labels. Because these labels are
rendered using dom-repeat templates, the re-rendering can occur in a
bottom-up direction. If the destination change is in a project that
allows different labels than the source change, this bottom-up re-render
may attempt to determine the description of a label which is not
available on the source change.

With this change, _computeValueTooltip is taught to fail gracefully when
asked to determine the description of a label score that is not
available on the destination change.

Bug: Issue 7288
Change-Id: I27e4044c51aa85c9fafa6c99161f82e2c99a246b

Former-commit-id: c5164c99227b2552a81a0e1e9b23df85e614cc80