Fix reference to preferred email select in gr-registration-dialog
In I31091ff247 the gr-select hybrid element was updated to a regular
Polymer component for forward compatibility, however, this broke the way
the dialog extracts its value when setting the preferred email address.
Because the dialog did not need the features provided by gr-select (i.e.
did not respond to change events) the gr-select is removed and the
vanilla select works as expected.

Bug: Issue 6867
Change-Id: Ic692995d5e9c4eb8271d5c97fa858e5e66718554

Former-commit-id: 96ae98def1c82af0e13880449d73ac3750aad986