Fix provisioning of ReviewDb in SystemGroupBackend.NameCheck
When the server is started, we don't have a request context which
provides an instance of ReviewDb. Therefore, we need to use the
SchemaFactory to obtain access to ReviewDb.

This issue was introduced by I52f8c66ec. Before that change, we also
used a SchemaFactory in GroupCacheImpl to access ReviewDb (probably
for exactly this reason). Unfortunately, our tests didn't catch this
issue. It only surfaces when GroupsIT#getSystemGroupByConfiguredName
is executed without any other tests of GroupsIT before it.

Change-Id: Ibdaea4bd99a7fb9aae2bdb73165a0c4e9df267c8

Former-commit-id: 5804c47566f0ba72276c304a99e14338a23b40c8