Implement custom dashboards
Custom dashboards are configured by query parameters. Any name/value
pair (with the exception of the name "title") configures a section and
its query. Sections are displayed in the order they appear in the query
string.

If a custom query contains any "${user}" placeholders, they will be
replaced with "self".

Bug: Issue 3941
Change-Id: If11aa7b3bdd1973e1ff36011a914096bde2fff6b

Former-commit-id: aa501d05d201aa9e9308a2f583abc7fd17ba00b7