Make ChangeControl package-private
This commit migrates all remaining callers of ChangeControl to
PermissionBackend and removes calls where unnecessary. These came out of
the drafts deprecation in Icfcb34efe1. With drafts, it was required to
check visibility on a patchset level. Some of there calls were simply
migrated to change level checks in Icfcb34efe1 and can now just be
removed. The rest of callers was migrated.

Change-Id: Ib5c9363778a678d3198656cc29c35848d5bd1bbd

Former-commit-id: 1cead4f79509f703df90cd0659cbbae840b710f1