Fall-back to loading groups from ReviewDb if group index is not available
In Gerrit slaves indexes are not available and hence trying to query
groups from the group index fails. This prevents loading of accounts via
the account cache. As result of this the account cache returns inactive
account instances to represent the missing accounts. This prevents users
from authenticating with SSH and SSH requests are rejected with
"Permission denied (publickey)."

To fix this fall-back to loading groups from ReviewDb if the group index
is not available. This reverts change I86f85b6e41 partially.

Bug: Issue 7340
Change-Id: I196c313f75f0a2f2bcb97445347bfe7113283198
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: b70b92ca57b1bfc2e7361e973c7ff5d041293ff8