Fix migration of mute stars
Schema migration 161 which should migrate 'mute' stars to 'reviewed'
stars was pretty broken:

1. It didn't allow non-fast-forward updates.
2. It didn't set the correct oldId in the ReceiveCommands.
3. It checked for 'mute' labels, but actually the label is
   'mute/<patch-set-number>'.

As result of this the schema migration didn't update any star refs.

There are 2 possibilities to fix this:
1. Fix schema migration 161 (implemented in this change):
   - The migration will not be done for sites that have already upgraded
     to 2.15-rc0. This is likely okay since:
     * the mute functionality didn't exist in 2.14 and hence for sites
       that upgraded from 2.14 there was nothing to migrate
     * it's unlikely that production sites were upgraded to the first
       release candidate of a new release
     * the mute labels are no critical data and are quickly outdated by
       uploading new patch sets
2. Add a new schema migration that does the correct thing:
   - We normally can't add schema migrations in stable branches because
     they would conflict with schema migrations in master. But in this
     case we don't have any new schema migration in master yet, so we
     could do this instead.

The replacement of Mute/Unmute with Mark Reviewed/Unreviewed was
implemented in several steps (first add support for Mark
Reviewed/Unreviewed, then removal of Mute/Unmute). For the unlikely case
that sites run a version were both Mute/Unmute and Mark
Reviewed/Unreviewed was possible the schema migration now also handles
the case where a patch set has the 'mute' label in addition to a
'reviewed' or 'unreviewed' label. In this case we trust the new labels
and the 'mute' label is just dropped.

Change-Id: I0941b5584ae6be3c1bbb3b72594cb34e67e8a48c
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: ed0b6dadf87010ac17c8156d530003280c7644c6