Use regex for query route pattern
Query URLs in GWT permitted un-encoded slashes in project operators,
however the route pattern for queries in PolyGerrit used the string
style of definition which would not permit slashes in params. With this
change, the two patterns for searches are unified into a single regex
style of definition which permits such slashes.

Bug: Issue 7306
Bug: Issue 9231
Change-Id: I597d9f0f3d1cb747998bb4213e21a663944f9ce6

Former-commit-id: e0b0449cce31230177c16899fa68ed8a40069134