Add extension point to allow custom merge super set computations
At Google we have performance issues with the Submitted Together REST
endpoint when topics span hundreds of projects. This extension point
allows us to distribute the load to compute the merge super set across
multiple server nodes and thus enhance the performance of the Submitted
Together REST endpoint.

In the multi-master setup at Google the load balancer distributes
requests to different nodes based on the project in the URL. This means
not all nodes have to load all repositories into cache which improves
performance a lot. However we still have performance problems with the
Submitted Together REST endpoint and for large topics which span several
hundreds of projects loading the list of changes that should be
submitted together can take several minutes. The problem is that the
request to the Submitted Together REST endpoint is not specific to one
project, but to include the predecessor change we have to open the
repository for all projects that are touched in the topics. This means
the one node node that handles the request has to open several hundreds
of repositories which is slow and also polutes the repository caches. To
fix this we want to implement the new extension point in such a way that
it distributes the load over multiple nodes that each have the necessary
repositories already cached. These nodes can do the computation in
parallel which should speed up the operation further.

Change-Id: I814498a7526cd5d818efed6cc000c03fc9bc5a09
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 19e89c86f00a84a4e1504c6374398a43d707930a